Gunnar: compatibility issues with sun compilers
git-svn-id: https://www.nick-ux.org/diagham/svn/trunk@1574 b9410c57-6ef9-0310-98bc-a73ba792996a

+++ FQHE/src/HilbertSpace/BosonOnLatticeGeneric.cc
@@ -386,7 +386,7 @@ unsigned long BosonOnLatticeGeneric::Ad (unsigned long state, int q, double& coe
   else
     {
       ++this->TemporaryState[q];
-      coefficient = sqrt(this->TemporaryState[q]);
+      coefficient = sqrt((double)this->TemporaryState[q]);
     }
   return this->BosonToFermion(this->TemporaryState, this->TemporaryStateHighestBit);
 }
